requirements.txt
streamlit
streamlit-option-menu
Pillow
google-generativeai